refactor(reply): relocate orchestration modules
